topic-ms-00001: Geospatial 타입의 값을 Long에서 Double로 변경.

+++ common/gourmet-common-type/src/main/kotlin/com/gourmet/service/common/type/Geospatial.kt
@@ -2,7 +2,7 @@ package com.gourmet.service.common.type
 
 abstract class Geospatial<T : Geospatial<T>>(
     val dimension: Int = 0,
-    val coordinates: Array<Long>
+    val coordinates: Array<Double>
 ) {
     abstract fun calcDistance(other: Geospatial<T>) : T
     abstract operator fun plus(other: Geospatial<T>) : T
